chore: update build-tools to 27.0.3
Update build tools to 27.0.3.

With latest android runtime we use gradle 4.4 and it downloads build-tools-27.0.3 at some point. Then CLI try to use them, but license is not accepted. Result is failing android build.

This will install latest build tools (and accept the license), so users will not encounter the issue described above.